update lên cầu thang đã hoàn thiện cầu thang 1

+++ GameCastlevania/04-Collision/Stage2.h
@@ -70,6 +70,6 @@ public:
 		game = g;
 	}
 	int column, row;
-	int **location2;
+	float **location2;
 	~CStage2();
 };
